Factor out colormaps
