+++ program/colormap.h
@@ -1,0 +1,35 @@
+// Copyright 2022 Anton Korobeynikov
+
+// This file is part of Bandage
+
+// Bandage is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+//(at your option) any later version.
+
+// Bandage is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+
+// You should have received a copy of the GNU General Public License
+// along with Bandage.  If not, see <http://www.gnu.org/licenses/>.
+
+#pragma once
+
+#include <colormap/tinycolormap_fwd.hpp>
+
+#include <QString>
+#include <QColor>
+
+enum ColorMap : int {
+    Viridis = 0,
+    Parula, Heat, Jet, Turbo, Hot, Gray, Magma, Inferno, Plasma, Cividis, Github, Cubehelix
+};
+
+ColorMap colorMapFromName(const QString& name);
+tinycolormap::ColormapType colorMap(ColorMap colorMap);
+QString getColorMapName(ColorMap colorMap);
+
+std::vector<QColor> getPresetColours();
+QString getColourName(QColor colour);
